backend/requirements.txt
Django==4.0
gunicorn==20.1.0
python-decouple
python-dotenv
djangorestframework